Update Azure_Pass_HowTo-zho-HK
Adjusted the translation back to English for the commands.  In this scenario, LOD is looking for a command named [azure-pass] and [balance]. If translated, LOD does not see a literal command of this naming convention, causing the command to fail.  In addition to this, adjusted the powershell command to be Start-process as well.